Refactor keyboard UI code out of ContentView
