+++ ADHDCoach/ContentView.swift
@@ -3,21 +3,7 @@ import Combine
 import UIKit
 import AVFoundation
 
-// Debug outline mode enum for visual debugging
-enum DebugOutlineMode: String, CaseIterable {
-    case none = "None"
-    case scrollView = "ScrollView"
-    case keyboardAttachedView = "Keyboard View"
-    case messageList = "Message List"
-    case spacer = "Spacer"
-    case vStack = "VStack"
-    case zStack = "ZStack"
-    case textInput = "Text Input"
-    case safeArea = "Safe Area"
-}
-
-// Flag to enable/disable input view layout debugging logs
-var inputViewLayoutDebug = false
+// These components are defined within the project, so no need to import them
 
 struct ContentView: View {
     // MARK: - Environment Objects
@@ -704,25 +690,6 @@ struct MessageHeightPreferenceKey: PreferenceKey {
     }
 }
 
-// Extension to get keyboard modifier flags
-extension UIWindow {
-    var eventModifierFlags: UIKeyModifierFlags? {
-        // For when a hardware keyboard is attached to the device
-        if let event = UIApplication.shared.windows.first?.undocumentedCurrentEvent {
-            return event.modifierFlags
-        }
-        return nil
-    }
-    
-    private var undocumentedCurrentEvent: UIEvent? {
-        // Private API access to get current event - necessary for detecting modifier keys
-        let selector = NSSelectorFromString("_currentEvent")
-        if responds(to: selector) {
-            return perform(selector).takeUnretainedValue() as? UIEvent
-        }
-        return nil
-    }
-}
 
 
 // MARK: - Previews
@@ -763,776 +730,8 @@ extension UIWindow {
 
 
 
-// MARK: - KeyboardState
-class KeyboardState: ObservableObject {
-    // Default font and sizing for consistency - accessible to other components
-    var defaultFont: UIFont { UIFont.systemFont(ofSize: UIFont.systemFontSize * 1.25) }
-    var singleLineHeight: CGFloat { defaultFont.lineHeight + 16 } // Line height + padding
-    var buttonRowHeight: CGFloat = 54 // Height for the send button row
-    var defaultInputHeight: CGFloat { singleLineHeight + 16 + buttonRowHeight } // Add container padding + button row
-    
-    /// Current keyboard height when visible, or 0 when hidden
-    @Published var keyboardOffset: CGFloat = 0
-    
-    /// Whether the keyboard is currently visible
-    @Published var isKeyboardVisible: Bool = false
-    
-    /// Height of the input view component
-    @Published var inputViewHeight: CGFloat = 0
-    
-    init() {
-        // Initialize with calculated default height
-        self.inputViewHeight = defaultInputHeight
-    }
-    
-    /// Updates keyboard state if there's an actual change to prevent unnecessary view updates
-    /// - Parameters:
-    ///   - visible: Whether keyboard is visible
-    ///   - height: Height of keyboard in points
-    func setKeyboardVisible(_ visible: Bool, height: CGFloat) {
-        // Always update if visible state changes
-        let visibilityChanged = isKeyboardVisible != visible
-        // For height changes, only check when keyboard is/remains visible
-        let heightChanged = visible && keyboardOffset != height
-        
-        if visibilityChanged {
-            withAnimation(.easeInOut(duration: 0.25)) {
-                isKeyboardVisible = visible
-                keyboardOffset = visible ? height : 0
-            }
-        } else if heightChanged {
-            withAnimation(.easeInOut(duration: 0.25)) {
-                keyboardOffset = height
-            }
-        }
-    }
-    
-    /// Returns the appropriate padding for the input view based on current keyboard state
-    /// - Parameters:
-    ///   - baseHeight: Default height to use when keyboard is hidden
-    ///   - safeAreaPadding: Additional padding to account for safe area
-    /// - Returns: The calculated padding value
-    func getInputViewPadding(baseHeight: CGFloat, safeAreaPadding: CGFloat) -> CGFloat {
-        // Get the correct height based on the text input size
-        let actualBaseHeight = inputViewHeight != defaultInputHeight ? inputViewHeight : baseHeight
-        
-        // When keyboard is visible, we need to account for both keyboard height AND the text input height difference
-        if isKeyboardVisible {
-            // Calculate height difference from default
-            let heightDifference = inputViewHeight - defaultInputHeight
-            
-            // With keyboard open, we need to ensure we have enough space for the text input AND button row
-            // The inputViewHeight includes text input + padding + button row
-            // When the keyboard is open, add the button row height to ensure proper space for both elements
-            return keyboardOffset + safeAreaPadding + buttonRowHeight + (heightDifference > 0 ? heightDifference : 0)
-        } else {
-            // When keyboard is hidden, just use the actual base height
-            return actualBaseHeight
-        }
-    }
-}
 
-// MARK: - TextInputView
-struct TextInputView: View {
-    // MARK: Properties
-    
-    // Input properties
-    @Binding var text: String
-    var onSend: () -> Void
-    
-    // Visual properties
-    var colorScheme: ColorScheme
-    var themeColor: Color
-    var isDisabled: Bool
-    var debugOutlineMode: DebugOutlineMode
-    
-    // Local state
-    @State private var isSending = false
-    @State private var textEditorHeight: CGFloat = 0 // Will be set to minHeight in onAppear
-    
-    // Constants 
-    private var defaultFont: UIFont { UIFont.systemFont(ofSize: UIFont.systemFontSize * 1.25) }
-    private var defaultFontSize: CGFloat { defaultFont.pointSize }
-    private var lineHeight: CGFloat { defaultFont.lineHeight }
-    private let maxLines: Int = 4
-    private var minHeight: CGFloat { lineHeight + 16 } // Single line + padding
-    
-    // Computed properties
-    private var isButtonDisabled: Bool {
-        isDisabled || text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty || isSending
-    }
-    
-    private var buttonColor: Color {
-        isButtonDisabled ? .gray : themeColor
-    }
-    
-    // Calculate max height based on max lines
-    private var maxHeight: CGFloat {
-        let padding: CGFloat = 16 // vertical padding
-        return (CGFloat(maxLines) * lineHeight) + padding
-    }
-    
-    // MARK: Body
-    var body: some View {
-        VStack(spacing: 8) {
-            // Text input field - using a multi-line editor, now full width
-            ZStack(alignment: .leading) {
-                // Placeholder text that shows when the text editor is empty
-                if text.isEmpty {
-                    Text("Message CosmiCoach")
-                        .foregroundColor(Color(.placeholderText))
-                        .padding(.horizontal, 6)
-                        .font(.system(size: defaultFontSize))
-                }
-                
-                // Actual text editor
-                MultilineTextField(text: $text, onSubmit: {
-                    // Submit on shift+return or command+return
-                    if !text.isEmpty {
-                        onSend()
-                    }
-                })
-                .frame(height: min(textEditorHeight, maxHeight))
-                .onAppear {
-                    // Initialize height with calculated minHeight
-                    textEditorHeight = minHeight
-                }
-                .onChange(of: text) { _, newText in
-                    // Calculate height based on text content
-                    let size = getTextSize(for: newText)
-                    let newHeight = min(max(size.height + 16, minHeight), maxHeight)
-                    
-                    // Only update if height actually changed
-                    if newHeight != textEditorHeight {
-                        textEditorHeight = newHeight
-                        
-                        // Calculate total height including padding, container, and button row
-                        let buttonRowHeight: CGFloat = 54 // Match the height in KeyboardState
-                        let totalHeight = newHeight + 16 + buttonRowHeight // Add padding for container + button row
-                        
-                        // Logging for debugging
-                        if inputViewLayoutDebug {
-                            print("Text height changed: \(textEditorHeight) → total: \(totalHeight)")
-                        }
-                        
-                        // Notify parent view controller of height change
-                        DispatchQueue.main.async {
-                            NotificationCenter.default.post(
-                                name: NSNotification.Name("InputViewHeightChanged"),
-                                object: nil,
-                                userInfo: ["height": totalHeight]
-                            )
-                        }
-                    }
-                }
-                .scrollContentBackground(.hidden)
-            }
-            .padding(.vertical, 4)
-            .padding(.horizontal, 8)
-            .border(debugOutlineMode == .textInput ? Color.pink : Color.clear, width: debugOutlineMode == .textInput ? 1 : 0)
-            .animation(nil, value: text) // Prevent animation during transitions
-            
-            // Send button row
-            HStack {
-                Spacer()
-                
-                Button {
-                    guard !isSending else { return }
-                    isSending = true
-                    
-                    // Reset text editor height before sending
-                    textEditorHeight = minHeight
-                    
-                    // Explicitly notify about height change to default
-                    if inputViewLayoutDebug {
-                        print("Resetting text height to default: \(minHeight)")
-                    }
-                    
-                    // Calculate total input view height (text + container padding + button row)
-                    let buttonRowHeight: CGFloat = 54 // Match the height in KeyboardState
-                    let totalHeight = minHeight + 16 + buttonRowHeight // Add padding for container + button row
-                    
-                    // Notify parent about height change
-                    NotificationCenter.default.post(
-                        name: NSNotification.Name("InputViewHeightChanged"),
-                        object: nil,
-                        userInfo: ["height": totalHeight]
-                    )
-                    
-                    // Call send after height is reset
-                    onSend()
-                    
-                    // Reset button state after a brief delay
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                        isSending = false
-                    }
-                } label: {
-                    ZStack {
-                        Circle()
-                            .foregroundColor(buttonColor)
-                            .frame(width: 34, height: 34)
-                        Image(systemName: "arrow.up")
-                            .font(.system(size: 20, weight: .semibold))
-                            .foregroundColor(isButtonDisabled ? Color(.systemBackground) : .white)
-                    }
-                }
-                .padding(.bottom, 4)
-                .padding(.horizontal, 2)
-                .disabled(isButtonDisabled)
-                .animation(.easeInOut(duration: 0.1), value: isButtonDisabled)
-            }
-            .padding(.horizontal, 8)
-        }
-        .padding(.horizontal, 4)
-        .padding(.vertical, 8)
-        .border(debugOutlineMode == .textInput ? Color.mint : Color.clear, width: 2)
-        .transaction { transaction in
-            transaction.animation = nil // Prevent position animations
-        }
-        .background(Color(.secondarySystemBackground))
-        .cornerRadius(25, corners: [.topLeft, .topRight])
-    }
-    
-    // Helper to calculate text size
-    private func getTextSize(for text: String) -> CGSize {
-        let attributes: [NSAttributedString.Key: Any] = [
-            .font: defaultFont
-        ]
-        
-        let width = UIScreen.main.bounds.width - 70 // Wider width since button is now in a separate row
-        let constraintRect = CGSize(width: width, height: .greatestFiniteMagnitude)
-        let boundingBox = text.boundingRect(
-            with: constraintRect,
-            options: .usesLineFragmentOrigin,
-            attributes: attributes,
-            context: nil
-        )
-        
-        return boundingBox.size
-    }
-}
 
-// MARK: - KeyboardAttachedView
-struct KeyboardAttachedView: UIViewControllerRepresentable {
-    // MARK: Properties
-    var keyboardManager: KeyboardManager
-    @Binding var text: String
-    var onSend: () -> Void
-    var colorScheme: ColorScheme
-    var themeColor: Color
-    var isDisabled: Bool
-    var debugOutlineMode: DebugOutlineMode
-    
-    // MARK: UIViewControllerRepresentable
-    func makeUIViewController(context: Context) -> KeyboardObservingViewController {
-        print("KeyboardAttachedView.makeUIViewController")
-        return KeyboardObservingViewController(
-            keyboardManager: keyboardManager,
-            text: $text,
-            onSend: onSend,
-            colorScheme: colorScheme,
-            themeColor: themeColor,
-            isDisabled: isDisabled,
-            debugOutlineMode: debugOutlineMode
-        )
-    }
-    
-    func updateUIViewController(_ uiViewController: KeyboardObservingViewController, context: Context) {
-        // Only update content if there are actual changes to avoid unnecessary view rebuilds
-        let textChanged = uiViewController.text.wrappedValue != text
-        let colorSchemeChanged = uiViewController.colorScheme != colorScheme
-        let themeColorChanged = uiViewController.themeColor != themeColor
-        let disabledStateChanged = uiViewController.isDisabled != isDisabled
-        let debugModeChanged = uiViewController.debugOutlineMode != debugOutlineMode
-        
-        if textChanged || colorSchemeChanged || themeColorChanged || disabledStateChanged || debugModeChanged {
-            uiViewController.updateContent(
-                text: text,
-                colorScheme: colorScheme,
-                themeColor: themeColor,
-                isDisabled: isDisabled,
-                debugOutlineMode: debugOutlineMode
-            )
-        }
-    }
-}
 
-// MARK: - MultilineTextField
-struct MultilineTextField: UIViewRepresentable {
-    @Binding var text: String
-    var onSubmit: () -> Void
-    
-    func makeUIView(context: Context) -> UITextView {
-        let textView = UITextView()
-        textView.delegate = context.coordinator
-        textView.font = UIFont.systemFont(ofSize: UIFont.systemFontSize * 1.25) // Use system font size with slight increase
-        textView.backgroundColor = .clear
-        textView.isScrollEnabled = true
-        textView.isEditable = true
-        textView.isUserInteractionEnabled = true
-        textView.autocapitalizationType = .sentences
-        textView.returnKeyType = .default // Use default return key
-        textView.keyboardType = .default
-        textView.textContainerInset = UIEdgeInsets(top: 8, left: 0, bottom: 8, right: 0)
-        textView.text = text
-        return textView
-    }
-    
-    func updateUIView(_ uiView: UITextView, context: Context) {
-        // Only update if text changed externally
-        if uiView.text != text {
-            uiView.text = text
-            
-            // When text is cleared, notify about height change
-            if text.isEmpty {
-                // Access defaultInputHeight via UIKit extension since we're in a UIViewRepresentable
-                let font = UIFont.systemFont(ofSize: UIFont.systemFontSize * 1.25)
-                let lineHeight = font.lineHeight
-                let buttonRowHeight: CGFloat = 54 // Match the height in KeyboardState
-                let defaultHeight = lineHeight + 16 + 16 + buttonRowHeight // Line height + inner padding + container padding + button row
-                
-                DispatchQueue.main.async {
-                    NotificationCenter.default.post(
-                        name: NSNotification.Name("InputViewHeightChanged"),
-                        object: nil,
-                        userInfo: ["height": defaultHeight]
-                    )
-                }
-            }
-        }
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-    
-    class Coordinator: NSObject, UITextViewDelegate {
-        var parent: MultilineTextField
-        
-        init(_ parent: MultilineTextField) {
-            self.parent = parent
-        }
-        
-        func textViewDidChange(_ textView: UITextView) {
-            parent.text = textView.text
-        }
-        
-        func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange, replacementText text: String) -> Bool {
-            // Check for special key combinations
-            let currentText = textView.text ?? ""
-            
-            // Current selection
-            let selectedRange = textView.selectedRange
-            
-            // Handle key combinations
-            if text == "\n" {
-                // Check if shift or command key is pressed
-                let modifierFlags = UIApplication.shared.windows.first?.windowScene?.keyWindow?.eventModifierFlags
-                
-                if modifierFlags?.contains(.shift) == true {
-                    // Shift+Return: submit message
-                    parent.onSubmit()
-                    return false
-                } else if modifierFlags?.contains(.command) == true {
-                    // Command+Return: submit message
-                    parent.onSubmit()
-                    return false
-                } else {
-                    // Normal Return: insert line break (default behavior)
-                    return true
-                }
-            }
-            
-            return true
-        }
-    }
-}
 
-// MARK: - KeyboardObservingViewController
-class KeyboardObservingViewController: UIViewController {
-    // MARK: Views
-    private var keyboardTrackingView = UIView()
-    private var safeAreaView = UIView()
-    private var inputHostView: UIHostingController<TextInputView>!
-    
-    // MARK: Constants
-    private let keyboardVisibilityThreshold: CGFloat = 100
-    
-    // MARK: Properties
-    internal var keyboardManager: KeyboardManager // For accessing keyboard properties
-    private var bottomConstraint: NSLayoutConstraint?
-    private var heightConstraint: NSLayoutConstraint?
-    private var _lastInputViewHeight: CGFloat = 0
-    internal var text: Binding<String> // Changed from private to internal
-    internal var onSend: () -> Void // Changed from private to internal
-    internal var colorScheme: ColorScheme // Changed from private to internal
-    internal var themeColor: Color // Changed from private to internal
-    internal var isDisabled: Bool // Changed from private to internal
-    internal var debugOutlineMode: DebugOutlineMode // Changed from private to internal
-    
-    // MARK: Lifecycle
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
-    
-    init(
-        keyboardManager: KeyboardManager,
-        text: Binding<String>,
-        onSend: @escaping () -> Void,
-        colorScheme: ColorScheme,
-        themeColor: Color,
-        isDisabled: Bool,
-        debugOutlineMode: DebugOutlineMode
-    ) {
-        self.keyboardManager = keyboardManager
-        self.text = text
-        self.onSend = onSend
-        self.colorScheme = colorScheme
-        self.themeColor = themeColor
-        self.isDisabled = isDisabled
-        self.debugOutlineMode = debugOutlineMode
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: View Lifecycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupViews()
-        setupKeyboardObservers()
-    }
-    
-    // MARK: View Setup
-    private func setupViews() {
-        setupKeyboardTrackingView()
-        setupSafeAreaView()
-        setupTextInputView()
-        updateDebugBorders()
-    }
-    
-    private func setupSafeAreaView() {
-        safeAreaView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(safeAreaView)
-        
-        NSLayoutConstraint.activate([
-            safeAreaView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
-            safeAreaView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
-            safeAreaView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            safeAreaView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor)
-        ])
-    }
-    
-    private func setupKeyboardTrackingView() {
-        keyboardTrackingView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(keyboardTrackingView)
-        keyboardTrackingView.backgroundColor = UIColor.secondarySystemBackground
-        
-        NSLayoutConstraint.activate([
-            // Pin horizontally to view edges
-            keyboardTrackingView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            keyboardTrackingView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            // Track keyboard vertically
-            keyboardTrackingView.topAnchor.constraint(equalTo: view.keyboardLayoutGuide.topAnchor),
-            keyboardTrackingView.bottomAnchor.constraint(equalTo: view.keyboardLayoutGuide.bottomAnchor)
-        ])
-    }
-    
-    private func setupTextInputView() {
-        // Create SwiftUI view
-        let textView = createTextInputView()
-        inputHostView = UIHostingController(rootView: textView)
-        
-        // Add hosting controller as child
-        addChild(inputHostView)
-        inputHostView.view.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(inputHostView.view)
-        inputHostView.didMove(toParent: self)
-        
-        // Set up constraints
-        NSLayoutConstraint.activate([
-            inputHostView.view.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            inputHostView.view.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-        ])
-        
-        // Create height constraint that we can update later - use keyboard manager's default height
-        heightConstraint = inputHostView.view.heightAnchor.constraint(equalToConstant: keyboardManager.defaultInputHeight)
-        heightConstraint?.isActive = true
-        
-        // Attach to keyboard
-        bottomConstraint = inputHostView.view.bottomAnchor.constraint(equalTo: view.keyboardLayoutGuide.topAnchor)
-        bottomConstraint?.isActive = true
-        
-        // Setup notification observer for input view height changes
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(handleInputViewHeightChange),
-            name: NSNotification.Name("InputViewHeightChanged"),
-            object: nil
-        )
-    }
-    
-    private func createTextInputView() -> TextInputView {
-        return TextInputView(
-            text: text,
-            onSend: onSend,
-            colorScheme: colorScheme,
-            themeColor: themeColor,
-            isDisabled: isDisabled,
-            debugOutlineMode: debugOutlineMode
-        )
-    }
-    
-    // MARK: Content Updates
-    func updateContent(
-        text: String,
-        colorScheme: ColorScheme,
-        themeColor: Color,
-        isDisabled: Bool,
-        debugOutlineMode: DebugOutlineMode
-    ) {
-        // Check for changes that would require us to update the view
-        let textChanged = self.text.wrappedValue != text
-        let themeColorChanged = self.themeColor != themeColor
-        let disabledStateChanged = self.isDisabled != isDisabled
-        let debugModeChanged = self.debugOutlineMode != debugOutlineMode
-        let colorSchemeChanged = self.colorScheme != colorScheme
-        let visualPropertiesChanged = themeColorChanged || disabledStateChanged || debugModeChanged || colorSchemeChanged
-        
-        // Skip full update if nothing has changed
-        if !textChanged && !visualPropertiesChanged {
-            return
-        }
-        
-        print("KeyboardObservingViewController.updateContent - updating \(textChanged ? "text" : "") \(visualPropertiesChanged ? "visuals" : "")")
-        
-        // Update text (without animation if clearing)
-        if textChanged && text.isEmpty {
-            CATransaction.begin()
-            CATransaction.setDisableActions(true)
-            self.text.wrappedValue = text
-            CATransaction.commit()
-        } else if textChanged {
-            self.text.wrappedValue = text
-        }
-        
-        // Update other properties
-        if visualPropertiesChanged {
-            self.colorScheme = colorScheme
-            self.themeColor = themeColor
-            self.isDisabled = isDisabled
-            self.debugOutlineMode = debugOutlineMode
-            
-            // Update SwiftUI view only if visual properties changed
-            inputHostView.rootView = createTextInputView()
-            updateSwiftUIViewPosition()
-        }
-        
-        // Update debug visualization
-        if debugModeChanged {
-            updateDebugBorders()
-            
-            // Update view order based on debug mode
-            if debugOutlineMode == .safeArea {
-                view.bringSubviewToFront(safeAreaView)
-            } else if debugOutlineMode == .keyboardAttachedView {
-                view.bringSubviewToFront(keyboardTrackingView)
-            }
-            
-            // Always keep input view on top
-            if let hostView = inputHostView?.view {
-                view.bringSubviewToFront(hostView)
-            }
-        }
-    }
-    
-    // MARK: Debug Visualization
-    private func updateDebugBorders() {
-        let isKeyboardAttachedDebug = debugOutlineMode == .keyboardAttachedView
-        let isSafeAreaDebug = debugOutlineMode == .safeArea
-        let isTextInputDebug = debugOutlineMode == .textInput
-        
-        // Keyboard tracking view
-        keyboardTrackingView.layer.borderWidth = isKeyboardAttachedDebug ? 2 : 0
-        keyboardTrackingView.layer.borderColor = UIColor.systemBlue.cgColor
-        
-        // Safe area visualization
-        safeAreaView.layer.borderWidth = isSafeAreaDebug ? 2 : 0
-        safeAreaView.layer.borderColor = UIColor.systemGreen.cgColor
-        safeAreaView.backgroundColor = isSafeAreaDebug ? 
-            UIColor.systemGreen.withAlphaComponent(0.1) : .clear
-        
-        // Main controller view
-        view.layer.borderWidth = (isKeyboardAttachedDebug || isSafeAreaDebug) ? 1 : 0
-        view.layer.borderColor = UIColor.systemTeal.cgColor
-        
-        // Text input host view
-        if let hostView = inputHostView?.view {
-            hostView.layer.borderWidth = isTextInputDebug ? 2 : 0
-            hostView.layer.borderColor = UIColor.systemIndigo.cgColor
-        }
-    }
-    
-    // MARK: Keyboard Observation
-    private func setupKeyboardObservers() {
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(keyboardWillChangeFrame),
-            name: UIResponder.keyboardWillChangeFrameNotification,
-            object: nil
-        )
-        
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(keyboardWillHide),
-            name: UIResponder.keyboardWillHideNotification,
-            object: nil
-        )
-    }
-    
-    @objc func keyboardWillChangeFrame(_ notification: Notification) {
-        guard let keyboardFrame = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect,
-              let duration = notification.userInfo?[UIResponder.keyboardAnimationDurationUserInfoKey] as? Double,
-              let curve = notification.userInfo?[UIResponder.keyboardAnimationCurveUserInfoKey] as? NSNumber else {
-            return
-        }
-        
-        // Check keyboard visibility
-        let isVisible = keyboardFrame.minY < UIScreen.main.bounds.height
-        
-        // Update state without SwiftUI animation (since we'll do UIKit animation)
-        CATransaction.begin()
-        CATransaction.setDisableActions(true)
-        keyboardManager.setKeyboardVisible(isVisible, height: keyboardFrame.height)
-        CATransaction.commit()
-        
-        // Match keyboard animation exactly using UIKit
-        let curveValue = curve.uintValue
-        let animationOptions = UIView.AnimationOptions(rawValue: curveValue << 16)
-        
-        // Animate with matching curve
-        UIView.animate(withDuration: duration, delay: 0, options: [animationOptions, .beginFromCurrentState]) {
-            self.view.layoutIfNeeded()
-            self.updateSwiftUIViewPosition()
-        }
-    }
-    
-    private func updateSwiftUIViewPosition() {
-        // Force layout update
-        inputHostView.view.setNeedsLayout()
-        inputHostView.view.layoutIfNeeded()
-    }
-    
-    @objc func keyboardWillHide(_ notification: Notification) {
-        guard let duration = notification.userInfo?[UIResponder.keyboardAnimationDurationUserInfoKey] as? Double,
-              let curve = notification.userInfo?[UIResponder.keyboardAnimationCurveUserInfoKey] as? NSNumber else {
-            return
-        }
-        
-        // Update state without SwiftUI animation (since we'll do UIKit animation)
-        CATransaction.begin()
-        CATransaction.setDisableActions(true)
-        keyboardManager.setKeyboardVisible(false, height: 0)
-        CATransaction.commit()
-        
-        // Match keyboard animation exactly
-        let curveValue = curve.uintValue
-        let animationOptions = UIView.AnimationOptions(rawValue: curveValue << 16)
-        
-        // Animate with matching curve
-        UIView.animate(withDuration: duration, delay: 0, options: [animationOptions, .beginFromCurrentState]) {
-            self.view.layoutIfNeeded()
-            self.updateSwiftUIViewPosition()
-        }
-    }
-    
-    // MARK: Input View Height Handling
-    @objc func handleInputViewHeightChange(_ notification: Notification) {
-        guard let height = notification.userInfo?["height"] as? CGFloat else { return }
-        
-        print("📝 Height change notification received: \(height) (current: \(_lastInputViewHeight))")
-        
-        // Store the height in our backing variable
-        _lastInputViewHeight = height
-        
-        // Update the height constraint
-        heightConstraint?.constant = height
-        
-        // Notify parent view to update layout for the new height
-        NotificationCenter.default.post(
-            name: NSNotification.Name("KeyboardStateChanged"),
-            object: nil,
-            userInfo: ["height": height]
-        )
-        
-        // Update layout with animation
-        UIView.animate(withDuration: 0.2) {
-            self.view.layoutIfNeeded()
-        }
-        
-        // Make sure we update the parent ContentView's layout as well
-        DispatchQueue.main.async {
-            self.updateSwiftUIViewPosition()
-        }
-        
-        // Ensure the parent views are updated as well
-        // This is especially important after sending a message
-        if text.wrappedValue.isEmpty || height == keyboardManager.defaultInputHeight {
-            print("📝 Text is empty or height is default, forcing reset to defaults")
-            
-            // Force another update after a slight delay to ensure it takes effect
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                self.heightConstraint?.constant = self.keyboardManager.defaultInputHeight
-                
-                UIView.animate(withDuration: 0.2) {
-                    self.view.layoutIfNeeded()
-                }
-            }
-        }
-    }
-    
-    // MARK: Interactive Gesture Handling
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-        
-        // Update SwiftUI view layout
-        updateSwiftUIViewPosition()
-        
-        // Handle interactive keyboard dismissal when no animation is in progress
-        guard let window = view.window, UIView.inheritedAnimationDuration == 0 else { return }
-        updateKeyboardPositionDuringInteractiveGesture(in: window)
-    }
-    
-    private func updateKeyboardPositionDuringInteractiveGesture(in window: UIWindow) {
-        // Get keyboard position
-        let keyboardFrame = view.keyboardLayoutGuide.layoutFrame
-        let screenHeight = window.frame.height
-        
-        // Convert to window coordinates
-        let keyboardFrameInWindow = view.convert(keyboardFrame, to: window)
-        let keyboardTop = keyboardFrameInWindow.minY
-        
-        // Calculate visibility
-        let keyboardHeight = screenHeight - keyboardTop
-        let isVisible = keyboardTop < screenHeight && keyboardHeight > keyboardVisibilityThreshold
-        
-        // Check if update needed
-        let heightDifference = abs(keyboardManager.keyboardOffset - (isVisible ? keyboardHeight : 0))
-        let shouldUpdate = heightDifference > 1.0 || keyboardManager.isKeyboardVisible != isVisible
-        
-        if shouldUpdate {
-            // Update state without animation during interactive gesture
-            CATransaction.begin()
-            CATransaction.setDisableActions(true)
-            keyboardManager.setKeyboardVisible(isVisible, height: isVisible ? keyboardHeight : 0)
-            CATransaction.commit()
-            
-            // Update layout immediately
-            view.layoutIfNeeded()
-            updateSwiftUIViewPosition()
-        }
-    }
-}
