+++ ADHDCoach/DebugOutlineMode.swift
@@ -1,0 +1,14 @@
+import SwiftUI
+
+// Debug outline mode enum for visual debugging
+public enum DebugOutlineMode: String, CaseIterable {
+    case none = "None"
+    case scrollView = "ScrollView"
+    case keyboardAttachedView = "Keyboard View"
+    case messageList = "Message List"
+    case spacer = "Spacer"
+    case vStack = "VStack"
+    case zStack = "ZStack"
+    case textInput = "Text Input"
+    case safeArea = "Safe Area"
+}
